+++ opm/simulators/flow/OutputCompositionalModule.hpp
@@ -52,7 +52,8 @@
 #include <opm/output/eclipse/Inplace.hpp>
 
 #include <opm/simulators/flow/FlowBaseVanguard.hpp>
-#include <opm/simulators/flow/GenericOutputCompositionalModule.hpp>
+// #include <opm/simulators/flow/GenericOutputCompositionalModule.hpp>
+#include <opm/simulators/flow/GenericOutputBlackoilModule.hpp>
 
 #include <algorithm>
 #include <array>
@@ -86,7 +87,7 @@ class EcfvDiscretization;
  *        ECL binary format.
  */
 template <class TypeTag>
-class OutputCompositionalModule : public GenericOutputCompositionalModule<GetPropType<TypeTag, Properties::FluidSystem>>
+class OutputCompositionalModule : public GenericOutputBlackoilModule<GetPropType<TypeTag, Properties::FluidSystem>>
 {
     using Simulator = GetPropType<TypeTag, Properties::Simulator>;
     using Discretization = GetPropType<TypeTag, Properties::Discretization>;
@@ -100,7 +101,7 @@ class OutputCompositionalModule : public GenericOutputCompositionalModule<GetPro
     using GridView = GetPropType<TypeTag, Properties::GridView>;
     using Element = typename GridView::template Codim<0>::Entity;
     using ElementIterator = typename GridView::template Codim<0>::Iterator;
-    using BaseType = GenericOutputCompositionalModule<FluidSystem>;
+    using BaseType = GenericOutputBlackoilModule<FluidSystem>;
     using Indices = GetPropType<TypeTag, Properties::Indices>;
     using Dir = FaceDir::DirEnum;
